new JPA version refactoring update

+++ project/src/main/java/com/jobworld/project/dto/applyViewDto/UserResumeRecruitDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UserResumeRecruitDTO {
 	private Long recruit_id;
-	private int resume_id;
+	private Long resume_id;
 	private String resume_title;
 	private String user_email;
 	private String user_nm;
